Leonie: Shut up a warning.

+++ LEOInterpreter.c
@@ -99,7 +99,7 @@ void	LEODebugPrintInstr( LEOInstruction* instruction )
 
 void	LEODebugPrintInstructions( LEOInstruction instructions[], size_t numInstructions )
 {
-	printf( "%u INSTRUCTIONS:\n", numInstructions );
+	printf( "%u INSTRUCTIONS:\n", (unsigned int)numInstructions );
 	for( int x = 0; x < numInstructions; x++ )
 	{
 		printf( "    " );
